Move Cache things to foundation
