+++ packages/hydrogen/src/components/LocalizationProvider/LocalizationProvider.server.tsx
@@ -2,7 +2,7 @@ import React, {ReactNode} from 'react';
 import LocalizationClientProvider from './LocalizationClientProvider.client';
 import {useShop} from '../../foundation/useShop';
 import {useShopQuery} from '../../hooks/useShopQuery';
-import {CacheDays} from '../../framework/CachingStrategy';
+import {CacheDays} from '../../foundation/Cache/strategies';
 import {PreloadOptions} from '../../types';
 import {Country, CountryCode} from '../../storefront-api-types';
 import {useSession} from '../../foundation/useSession/useSession';
