+++ packages/hydrogen/src/foundation/HydrogenResponse/HydrogenResponse.server.ts
@@ -1,7 +1,4 @@
-import {
-  CacheSeconds,
-  generateCacheControlHeader,
-} from '../../framework/CachingStrategy';
+import {CacheSeconds, generateCacheControlHeader} from '../Cache/strategies';
 import type {CachingStrategy} from '../../types';
 import Redirect from '../Redirect/Redirect.client';
 import React from 'react';
